% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Une fonction qui calcule la distribution de l'âge des élus}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma}
}
\value{
l'âge des élus de la ville/ du département correspondant aux quantiles
}
\description{
Une fonction pour avoir la distribution de l'âge des élus (quantiles) dans une ville/département
}
\examples{
calcul_distribution_age(df_Nantes) # Devrait retourner un tableau avec les quantiles en ligne et la distribution de l'âge des élus sur la colonne
}
