% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nb_commune.R
\name{compter_nb_commune}
\alias{compter_nb_commune}
\title{Une fonction qui trouve le nombre de commune d'un département}
\usage{
compter_nb_commune(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma}
}
\value{
le nombre de commune du département
}
\description{
Une fonction pour calculer le nombre de commune dans un département
}
\examples{
compter_nb_commune(df_Loire_Atlantique) # Devrait retourner 207
}
