% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nb_elus.R
\name{compter_nb_elus}
\alias{compter_nb_elus}
\title{Une fonction qui compte le nombre d'élus}
\usage{
compter_nb_elus(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma décrit}
}
\value{
le nombre d'élus uniques
}
\description{
Une fonction pour compter le nombre d'élus uniques par ville/département
}
\examples{
compter_nb_elus(df_Nantes)
66
}
