% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Une fonction qui compte le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma décrit}
}
\value{
le nombre d'adjoints uniques
}
\description{
Une fonction pour compter le nombre d'adjoints uniques par ville/département
}
\examples{
compter_nombre_d_adjoints(df_Nantes) # Devrait retourner 26
}
