% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\name{creer_commune}
\alias{creer_commune}
\title{Fonction pour réer un objet de classe "commune"}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un dataframe avec les informations d'une commune}
}
\value{
Un objet de classe "commune"
}
\description{
Une fonction qui ajoute la classe "commune" à un dataframe
}
