% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Fonction pour réer un objet de classe "departement"}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un dataframe avec les informations d'un département}
}
\value{
Un objet de classe "departement"
}
\description{
Fonction pour réer un objet de classe "departement"
}
