% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_elus.R
\docType{data}
\name{donnees_elus}
\alias{donnees_elus}
\title{Données sur les élus municipaux}
\format{
Un data frame avec 489 972 lignes et 16 colonnes :
\describe{
\item{Code_du_département}{code du département}
\item{Libellé_du_département}{nom du département}
\item{Code_de_la_collectivité_à_statut_particulier}{code de la collectivité à statut particulier}
\item{Libellé_de_la_collectivité_à_statut_particulier}{nom de la collectivité à statut particulier}
\item{Code_de_la_commune}{code de la commune}
\item{Libellé_de_la_commune}{nom de la commune}
\item{Nom_de_l_élu}{nom de l'élu}
\item{Prénom_de_l_élu}{prénom de l'élu}
\item{Code_sexe}{"M" si l'élu est un homme, "F" si l'élu est une femme}
\item{Date_de_naissance}{date de naissance de l'élu au format "jj/mm/aaaa"}
\item{Code_de_la_catégorie_socio_professionnelle}{code correspondant à la catégorie socio-professionnelle de l'élu}
\item{Libellé_de_la_catégorie_socio_professionnelle}{nom de la catégorie socio-professionnelle de l'élu}
\item{Date_de_début_du_mandat}{date de début du mandat au format "jj/mm/aaaa"}
\item{Libellé_de_la_fonction}{nom de la fonction de l'élu}
\item{Date_de_début_de_la_fonction}{date de début de la fonction au format "jj/mm/aaaa"}
\item{Code_nationalité}{code indiquant la nationalité de l'élu ("FR" s'il est français)}
}
}
\source{
\href{https://www.data.gouv.fr/fr/datasets/repertoire-national-des-elus-1/#/resources/d5f400de-ae3f-4966-8cb6-a85c70c6c24a}{Répertoire National des Élus}
}
\usage{
data(donnees_elus)
}
\description{
Tableau contenant les élus des communes françaises
}
\keyword{datasets}
