% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Génère un rapport Quarto pour une commune ou un département}
\usage{
generer_rapport(commune, departement, output = "rapport.html")
}
\arguments{
\item{commune}{Code de la commune}

\item{departement}{Code du département}

\item{output}{Chemin du fichier de sortie du rapport}
}
\value{
Génère un rapport au format HTML
}
\description{
Une fonction qui génère un rapport Quarto au format html correspondant à la commune et au département choisis. Si un document avec le même nom et au même endroit existe déjà, au nom de l'ancien rapport sera ajouté "old" ainsi que la date à laquelle le rapport à été généré.
}
