% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plot.commune.R, R/plot.departement.R
\name{plot}
\alias{plot}
\alias{plot.commune}
\alias{plot.departement}
\title{Fonction pour produire un graphique par rapport à un dataframe selon sa classe}
\usage{
plot(df)

\method{plot}{commune}(x)

\method{plot}{departement}(x)
}
\arguments{
\item{df}{Un dataframe avec les données nécessaires pour le graphique}

\item{x}{Un dataframe de classe "departement".}
}
\value{
Un graphique, selon la classe de \code{df}

Un diagramme en barre horizontal, avec les codes professionnels des élus en ordonné, et le nombre d'élus sur l'axe des abcisses

Un diagramme en barre horizontal, avec les 10 codes professionnels des élus les plus représentés en ordonné, et le nombre d'élus sur l'axe des abcisses
}
\description{
Fonction générique qui retourne un graphique différent selon la classe du dataframe.

Graphique pour un dataframe de classe "commune".

Graphique pour un dataframe de classe "departement".
}
\seealso{
\code{\link{plot.commune}}, \code{\link{plot.departement}}
}
