% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Une fonction qui produit un graphique sur le nombre d'élus par code professionnel}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma}
}
\value{
un bar chart horizontal, avec le nombre d'élus sur l'axe des abcisses et le code professionnel sur l'axe des ordonnées
}
\description{
Une fonction pour avoir un graphique en barre horizontal montrant le nombre d'élus par code professionnel pour une commune/un département
}
\examples{
plot_code_professions(df_Nantes)
}
