% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R, R/summary.commune.R,
%   R/summary.departement.R
\name{summary}
\alias{summary}
\alias{summary.commune}
\alias{summary.departement}
\title{Fonction pour résumer des informations d'un dataframe selon sa classe}
\usage{
summary(df)

\method{summary}{commune}(x)

\method{summary}{departement}(x)
}
\arguments{
\item{df}{Un dataframe avec les données à résumer}

\item{x}{Un dataframe de classe "departement".}
}
\value{
Un résumé des données, selon la classe de \code{df}

Un résumé d'information sur une commune, avec le libellé de la commune, le nombre d'élus, la distribution de l'âge des élus, des informations sur l'élu le/la plus agé.e

Un résumé d'information sur un département, avec le libellé du département, le nombre de commune, le nombre d'élus, la distribution de l'âge des élus, des informations sur l'élu le/la plus agé.e et l'élu le/la plus jeune, leur commune, le nom de la commune avec la moyenne d'âge la plus faible et avec la moyenne d'âge la plus élevée, la distribution des âges pour ces communes
}
\description{
Fonction générique qui retourne un résumé d'informations différent selon la classe du dataframe.

Résumé des informations pour un dataframe de classe "commune".

Résumé des informations pour un dataframe de classe "departement".
}
\seealso{
\code{\link{summary.commune}}, \code{\link{summary.departement}}
}
