% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Une fonction qui trouve l'élu le plus agé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma}
}
\value{
le nom, le prénom, et la date de naissance de l'élu le plus agé de la commune/département
}
\description{
Une fonction pour trouver l'élu le plus agé dans une ville/département
}
\examples{
trouver_l_elu_le_plus_age(df_Nantes)
}
