% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_l_elu_le_plus_jeune.R
\name{trouver_l_elu_le_plus_jeune}
\alias{trouver_l_elu_le_plus_jeune}
\title{Une fonction qui trouve l'élu le plus jeune}
\usage{
trouver_l_elu_le_plus_jeune(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma}
}
\value{
le nom, le prénom, et l'âge de l'élu le plus jeune de la commune/département
}
\description{
Une fonction pour trouver l'élu le plus jeune dans une ville/département
}
\examples{
trouver_l_elu_le_plus_jeune(df_Nantes)
}
