% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_schema.R
\name{validate_schema}
\alias{validate_schema}
\title{Une fonction qui valide le schéma du df}
\usage{
validate_schema(df)
}
\arguments{
\item{df}{le dataframe doit contenir les colonnes correspondantes au schéma décrit}
}
\value{
rien mais stop la fonction si le df ne correspond pas
}
\description{
Une fonction pour valider le fait que le schéma du dataframe suive le schéma souhaité
}
